requirements.txt
anyio==4.4.0
bcrypt==4.2.0
beanie==1.26.0
cffi==1.17.0
click==8.1.7
cryptography==3.4.8
dnspython==2.6.1
fastapi==0.112.0
fastapi-jwt-auth==0.5.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.7
lazy-model==0.2.0
MarkupSafe==2.1.5
motor==3.3.1
numpy==2.0.1
pandas==2.0.1
passlib==1.7.4
pycparser==2.22
pydantic==1.10.7
PyJWT==1.7.1
pymongo==4.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.5
Werkzeug==2.3.4
